pyproject.toml
[project]
name = "deploy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.10.0",
    "ollama>=0.6.1",
    "python-dotenv>=1.2.2",
]
